classe are now seed files. Need to put seeReq function in pageviews file
