+++ app.py
@@ -11,9 +11,6 @@
 
 import streamlit as st
 import pageviews as view
-import NewClasses as classes
-from NewClasses import kid_list
-from NewClasses import Kids
 
 st.title("Toys for Tots!")
 st.text("Here you can make donations to ensure children have Merry Christmases!")
@@ -31,4 +28,4 @@ elif options == 'Make Donation':
 elif options == 'Request Donation':
     view.makeReq()
 elif options == 'See Requested Donations':
-    classes.seeReq(Kids.parentName, Kids.kidName, Kids.age, Kids.toy)
+    view.seeReq()
